Update AStarPlanner.cpp
If x==0 then x-OBSTACLE_CLEARANCE=-1  which get's converted to the largest possible unsigned int, then this number gets compared to 0, and then the largest possible unsigned int is chosen as the maximum which then gets converted back to an int, making x_range_min -1. Also, x_range_max and z_range_max are closed bounds so their upper bound has to be equal to the largest possible cell index.